Fix client side errors from deleting tag, refactor removeTag actions and reducers
