+++ client/app/src/actions/tagActions.jsx
@@ -19,11 +19,3 @@ export const fetchTagsAction = () => {
     payload: request,
   };
 };
-
-export const removeTagAction = (tag) => {
-  const request = axios.delete(endpoints.card_tags + '/' + tag.cardTagId);
-  return {
-    type: 'REMOVE_TAG',
-    payload: request,
-  };
-};
